requirements.txt
pybit
pandas
numpy
python-dotenv
pydantic
loguru
websocket-client
